+++ bundles/org.corpus_tools.hexatomic.formats/src/main/java/org/corpus_tools/hexatomic/formats/importer/ImporterSelectionPage.java
@@ -59,6 +59,7 @@ public class ImporterSelectionPage extends CorpusFormatSelectionPage<ImportForma
     // Initialize selection to default
     btnExb.setSelection(false);
     btnPaulaXml.setSelection(false);
+    btnGraphAnno.setSelection(false);
     btnTxt.setSelection(false);
     Optional<Pepper> pepper = Activator.getPepper();
     setPageComplete(false);
@@ -78,6 +79,9 @@ public class ImporterSelectionPage extends CorpusFormatSelectionPage<ImportForma
             setPageComplete(true);
           } else if (format.get() == ImportFormat.PAULA) {
             btnPaulaXml.setSelection(true);
+            setPageComplete(true);
+          } else if (format.get() == ImportFormat.GRAPHANNO) {
+            btnGraphAnno.setSelection(true);
             setPageComplete(true);
           } else if (format.get() == ImportFormat.TXT) {
             btnTxt.setSelection(true);
@@ -101,6 +105,8 @@ public class ImporterSelectionPage extends CorpusFormatSelectionPage<ImportForma
       return Optional.of(ImportFormat.EXB);
     } else if (btnPaulaXml.getSelection()) {
       return Optional.of(ImportFormat.PAULA);
+    } else if (btnGraphAnno.getSelection()) {
+      return Optional.of(ImportFormat.GRAPHANNO);
     } else if (btnTxt.getSelection()) {
       return Optional.of(ImportFormat.TXT);
     }